fix: use AtomicUsize for cross-platform compatibility
- Use AtomicUsize instead of AtomicU64 for better portability
- Fix thiserror import

+++ shadowquic/src/quic/mod.rs
@@ -1,9 +1,10 @@
 use std::net::{SocketAddr, UdpSocket};
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::AtomicUsize;
 
 use crate::error::SResult;
 use async_trait::async_trait;
 use bytes::Bytes;
+use std::sync::atomic::Ordering;
 use thiserror::Error;
 use tokio::io::{AsyncRead, AsyncWrite};
 
@@ -18,23 +19,23 @@ pub const MAX_CONCURRENT_BIDI_STREAMS: u32 = 1000;
 pub const MAX_CONCURRENT_UNI_STREAMS: u32 = 1000;
 
 pub struct Stats {
-    pub total_connections: AtomicU64,
-    pub active_connections: AtomicU64,
-    pub total_packets_sent: AtomicU64,
-    pub total_packets_received: AtomicU64,
-    pub total_bytes_sent: AtomicU64,
-    pub total_bytes_received: AtomicU64,
+    pub total_connections: AtomicUsize,
+    pub active_connections: AtomicUsize,
+    pub total_packets_sent: AtomicUsize,
+    pub total_packets_received: AtomicUsize,
+    pub total_bytes_sent: AtomicUsize,
+    pub total_bytes_received: AtomicUsize,
 }
 
 impl Stats {
     pub const fn new() -> Self {
         Self {
-            total_connections: AtomicU64::new(0),
-            active_connections: AtomicU64::new(0),
-            total_packets_sent: AtomicU64::new(0),
-            total_packets_received: AtomicU64::new(0),
-            total_bytes_sent: AtomicU64::new(0),
-            total_bytes_received: AtomicU64::new(0),
+            total_connections: AtomicUsize::new(0),
+            active_connections: AtomicUsize::new(0),
+            total_packets_sent: AtomicUsize::new(0),
+            total_packets_received: AtomicUsize::new(0),
+            total_bytes_sent: AtomicUsize::new(0),
+            total_bytes_received: AtomicUsize::new(0),
         }
     }
 }
@@ -57,12 +58,12 @@ impl Stats {
 
     pub fn packet_sent(&self, size: usize) {
         self.total_packets_sent.fetch_add(1, Ordering::Relaxed);
-        self.total_bytes_sent.fetch_add(size as u64, Ordering::Relaxed);
+        self.total_bytes_sent.fetch_add(size, Ordering::Relaxed);
     }
 
     pub fn packet_received(&self, size: usize) {
         self.total_packets_received.fetch_add(1, Ordering::Relaxed);
-        self.total_bytes_received.fetch_add(size as u64, Ordering::Relaxed);
+        self.total_bytes_received.fetch_add(size, Ordering::Relaxed);
     }
 }
